deleted user records

+++ scripts/seed.ts
@@ -22,8 +22,8 @@ async function addCourses() {
         { name: "Physical Education and Health" },
         { name: "Economy" },
         { name: "Consumer Rights" },
-        { name: "Education for Work Financial and Tax Education" }
-      ]
+        { name: "Education for Work Financial and Tax Education" },
+      ],
     });
 
     console.log("Success");
@@ -35,116 +35,109 @@ async function addCourses() {
 }
 
 async function addAcademyLevels() {
-    try {
-      await database.academy.createMany({
-        data: [
-            { name: "Initial Education" },
-            { name: "Primary Education" },
-            { name: "Secondary Education" },
-        ]
-      })
-        
+  try {
+    await database.academy.createMany({
+      data: [
+        { name: "Initial Education" },
+        { name: "Primary Education" },
+        { name: "Secondary Education" },
+      ],
+    });
+
     console.log("Success");
-    } catch (error) {
-        
-        console.log("Error seeding the database academy Levels", error);
-    } finally {
-      await database.$disconnect();
-    }
+  } catch (error) {
+    console.log("Error seeding the database academy Levels", error);
+  } finally {
+    await database.$disconnect();
+  }
 }
 
 async function addEducationLevels() {
-    try {
-      await database.educationLevel.createMany({
-        data: [
-            { name: "Initial (3,4,5 years)" },
-            { name: "Primary(1st - 6th grades)" },
-            { name: "Secondary (1st - 5th grades)" },
-        ]
-      })
-        
+  try {
+    await database.educationLevel.createMany({
+      data: [
+        { name: "Initial (3,4,5 years)" },
+        { name: "Primary(1st - 6th grades)" },
+        { name: "Secondary (1st - 5th grades)" },
+      ],
+    });
+
     console.log("Success");
-    } catch (error) {
-        
-        console.log("Error seeding the database education Levels", error);
-    } finally {
-      await database.$disconnect();
-    }
+  } catch (error) {
+    console.log("Error seeding the database education Levels", error);
+  } finally {
+    await database.$disconnect();
+  }
 }
 
 async function addAverageAges() {
-    try {
-      await database.averageAge.createMany({
-        data: [
-            { name: "3" },
-            { name: "4" },
-            { name: "5" },
-            { name: "6" },
-            { name: "7" },
-            { name: "8" },
-            { name: "9" },
-            { name: "10" },
-            { name: "11" },
-            { name: "12" },
-            { name: "13" },
-            { name: "14" },
-            { name: "15" },
-            { name: "16" },
-            { name: "17" },
-            { name: "18" },
-        ]
-      })
-        
+  try {
+    await database.averageAge.createMany({
+      data: [
+        { name: "3" },
+        { name: "4" },
+        { name: "5" },
+        { name: "6" },
+        { name: "7" },
+        { name: "8" },
+        { name: "9" },
+        { name: "10" },
+        { name: "11" },
+        { name: "12" },
+        { name: "13" },
+        { name: "14" },
+        { name: "15" },
+        { name: "16" },
+        { name: "17" },
+        { name: "18" },
+      ],
+    });
+
     console.log("Success");
-    } catch (error) {
-        
-        console.log("Error seeding the database average age", error);
-    } finally {
-      await database.$disconnect();
-    }
+  } catch (error) {
+    console.log("Error seeding the database average age", error);
+  } finally {
+    await database.$disconnect();
+  }
 }
-
-
 
 async function addDepatments() {
   try {
     await database.department.createMany({
-      data : [
-        {"name": "Amazon"},
-         {"name": "Ancash"},
-        {"name": "Apurimac"},
-         {"name": "Arequipa"},
-        {"name": "Ayacucho"},
-         {"name": "Cajamarca"},
-       {"name": "Callao"},
-        {"name": "Cusco"},
-        {"name": "Huancavelica"},
-         {"name": "Huanuco"},
-        {"name": "Ica"},
-        {"name": "Junin"},
-        {"name": "Freedom"},
-        {"name": "Lambayeque"},
-        {"name": "Lime"},
-         {"name": "Loreto"},
-         {"name": "Mother of God"},
-         {"name": "Moquegua"},
-        {"name": "Pasco"},
-        {"name": "Piura"},
-        {"name": "Fist"},
-        {"name": "San Martin"},
-        {"name": "Tacna"},
-        {"name": "Tumbes"},
-        {"name": "Ucayali"}
-      ]
-    })
-    
+      data: [
+        { name: "Amazon" },
+        { name: "Ancash" },
+        { name: "Apurimac" },
+        { name: "Arequipa" },
+        { name: "Ayacucho" },
+        { name: "Cajamarca" },
+        { name: "Callao" },
+        { name: "Cusco" },
+        { name: "Huancavelica" },
+        { name: "Huanuco" },
+        { name: "Ica" },
+        { name: "Junin" },
+        { name: "Freedom" },
+        { name: "Lambayeque" },
+        { name: "Lime" },
+        { name: "Loreto" },
+        { name: "Mother of God" },
+        { name: "Moquegua" },
+        { name: "Pasco" },
+        { name: "Piura" },
+        { name: "Fist" },
+        { name: "San Martin" },
+        { name: "Tacna" },
+        { name: "Tumbes" },
+        { name: "Ucayali" },
+      ],
+    });
   } catch (error) {
-    console.log("An Error occured while adding departments")
-  } finally{
-    database.$disconnect()
+    console.log("An Error occured while adding departments");
+  } finally {
+    database.$disconnect();
   }
 }
-
 
 //EXPERIENCE MODEL SEEDING
 
@@ -152,16 +145,15 @@ async function addExperienceEducationLevels() {
   try {
     await database.educationLevel__experience.createMany({
       data: [
-          { name: "Initial (3,4,5 years)" },
-          { name: "Primary(1st - 6th grades)" },
-          { name: "Secondary (1st - 5th grades)" },
-      ]
-    })
-      
-  console.log("Success");
+        { name: "Initial (3,4,5 years)" },
+        { name: "Primary(1st - 6th grades)" },
+        { name: "Secondary (1st - 5th grades)" },
+      ],
+    });
+
+    console.log("Success");
   } catch (error) {
-      
-      console.log("Error seeding the database education Levels", error);
+    console.log("Error seeding the database education Levels", error);
   } finally {
     await database.$disconnect();
   }
@@ -171,33 +163,31 @@ async function addExperienceAverageAge() {
   try {
     await database.averageAge_experience.createMany({
       data: [
-          { name: "3" },
-          { name: "4" },
-          { name: "5" },
-          { name: "6" },
-          { name: "7" },
-          { name: "8" },
-          { name: "9" },
-          { name: "10" },
-          { name: "11" },
-          { name: "12" },
-          { name: "13" },
-          { name: "14" },
-          { name: "15" },
-          { name: "16" },
-          { name: "17" },
-          { name: "18" },
-      ]
-    })
-      
-  console.log("Success");
+        { name: "3" },
+        { name: "4" },
+        { name: "5" },
+        { name: "6" },
+        { name: "7" },
+        { name: "8" },
+        { name: "9" },
+        { name: "10" },
+        { name: "11" },
+        { name: "12" },
+        { name: "13" },
+        { name: "14" },
+        { name: "15" },
+        { name: "16" },
+        { name: "17" },
+        { name: "18" },
+      ],
+    });
+
+    console.log("Success");
   } catch (error) {
-      
-      console.log("Error seeding the database average age", error);
+    console.log("Error seeding the database average age", error);
   } finally {
     await database.$disconnect();
   }
-  
 }
 async function addExperienceCourseName() {
   try {
@@ -219,8 +209,8 @@ async function addExperienceCourseName() {
         { name: "Physical Education and Health" },
         { name: "Economy" },
         { name: "Consumer Rights" },
-        { name: "Education for Work Financial and Tax Education" }
-      ]
+        { name: "Education for Work Financial and Tax Education" },
+      ],
     });
 
     console.log("Success");
@@ -235,16 +225,15 @@ async function addExperienceAcademicLevel() {
   try {
     await database.academicLevel_experience.createMany({
       data: [
-          { name: "Initial Education" },
-          { name: "Primary Education" },
-          { name: "Secondary Education" },
-      ]
-    })
-      
-  console.log("Success");
+        { name: "Initial Education" },
+        { name: "Primary Education" },
+        { name: "Secondary Education" },
+      ],
+    });
+
+    console.log("Success");
   } catch (error) {
-      
-      console.log("Error seeding the database academy Levels", error);
+    console.log("Error seeding the database academy Levels", error);
   } finally {
     await database.$disconnect();
   }
@@ -252,21 +241,19 @@ async function addExperienceAcademicLevel() {
 
 async function deleteAllGaps() {
   try {
-      // Delete all gaps
-      await database.gAP.deleteMany({});
-      
-      console.log("All gaps deleted successfully");
+    const userId = "user_2aJRcHa8NsD7BrNqzjq6yWiKxbi";
+    // Delete all gaps
+    await database.gAP.deleteMany({ where: { userid: userId } });
+
+    console.log("All gaps deleted successfully");
   } catch (error) {
-      console.error("Error deleting gaps:", error);
+    console.error("Error deleting gaps:", error);
   } finally {
-      await database.$disconnect();
+    await database.$disconnect();
   }
 }
 
-// deleteAllGaps();
-
-
-
+deleteAllGaps();
 
 //GAPS
 // addEducationLevels()
@@ -277,7 +264,7 @@ async function deleteAllGaps() {
 
 //EXPERIENCE
 
-addExperienceCourseName()
-addExperienceAcademicLevel()
-addExperienceAverageAge()
-addExperienceEducationLevels()
+// addExperienceCourseName()
+// addExperienceAcademicLevel()
+// addExperienceAverageAge()
+// addExperienceEducationLevels()
